l3-adk/pyproject.toml
[project]
name = "l3-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.22.1",
]
